app/requirements.txt
rapidocr-onnxruntime 
pypdf 
langchain-text-splitters 
langchain-community 
langgraph 
langchain-openai 
langchain-core
langchain-chroma
fastapi
uvicorn
python-dotenv
google-auth==2.17.3
requests==2.28.2
asyncio